requirements.txt
flask>=3.0
pandas>=2.0
numpy>=1.24
pymysql>=1.1
requests>=2.31
python-dotenv>=1.0
